+++ database/migrations/2023_04_24_193322_create_report_rits_table.php
@@ -13,9 +13,9 @@ return new class extends Migration
     {
         Schema::create('report_rits', function (Blueprint $table) {
             $table->id();
-            $table->integer("tonnage_left");
-            $table->integer("tonnage_sold");
-            $table->integer("total_tonnage_sold");
+            $table->double("tonnage_left");
+            $table->double("tonnage_sold");
+            $table->double("total_tonnage_sold");
             $table->timestamps();
         });
     }
